Fix: setup signal handlers in interpreted code
The interpreter won't receive signals anymore, since the interpreted
code will receive them, but the interpreter's signal loop fiber won't be
resumed and it won't handle signals anymore.

Unblocks the default handler for SIGCHLD into the interpreted code, so
Process.run and waiting for a sub-process works just as expected.