Avoid flush in finalizers for `Socket` and `IO::FileDescriptor` (#14882)
Trying to flush means that it can try to write, hence call into the event loop, that may have to wait for the fd to be writable, which means calling into `epoll_wait`. This all while the world is stopped during a garbage collection run. The event loop implementation may need to allocate, or we get an error and try to raise an exception that will also try to allocate memory... again during a GC collection.

For `Socket` this change has little impact becase `sync` is true by default which means every byte is sent directly without buffering.
`IO::FileDescriptor` (and `File`) is buffered by default, so this may lead to loss of data if you don't properly close the file descriptor after use.