Experiment: single EventLoop for the process
This allows to keep a file descriptor into the evloop for its whole
lifetime (from open to close) instead of adding it every time it would
block and removing it as soon as it unblocks. This brings over 20%
performance improvement on a simple HTTP/1.1 server (with keepalive).

Among the advantages: this allows to remove the global mutex around
handling IO events and instead have an almost never contended lock
around the reader or the writer waiting lists for each IO. We don't even
have to keep a global list of events (epoll and kqueue will do it).

The drawback is that the preview MT scheduler isn't compatible with this
scheme.