Fix: Process.run hangs forever
This was caused by the signal pipe being reset before we reset the event
loop after fork and before exec. This led to remove the signal pipe fds
from the parent epoll/kqueue instance (and add the new fds to the
parent) which led the parent to not be capable to handle signals
anymore (and to receive signals for the child).