Sleep and select timeout don't need to handle nilables
Sleep and registering a select action are always setting a Time::Span
and don't need to deal with a nilable.

The exception is IO::Evented that keeps a cache of libevent events and
must be able to remove the timeout in case @read_timeout would have been
set to nil before a second wait on read.