Fix: `Crystal::SpinLock` doesn't need to be allocated on the HEAP (#14972)
The abstraction is a mere abstraction over an atomic integer and the
object itself are only ever used internally of other objects, with the
exception of Channel where the code explicitely accesses the ivar
directly (thus not making copies).

We can avoid a HEAP allocation everywhere we use them (i.e. in lots of places).