Don't involve evloop after fork in System::Process.spawn (UNIX) (#14974)
Refactors spawning child processes on UNIX that relies on fork/exec to not involve the event loop after fork and before exec. We still continue to rely on the eventloop in the parent process, of course.

* Extract Crystal::System::FileDescriptor.system_pipe (UNIX)
* Fix: avoid evloop after fork to report failures in Process.spawn (UNIX)
* Fix: don't involve evloop in System::Process.reopen_io (UNIX)

This is called after fork before exec to reopen the stdio. We can
leverage some abstractions (set blocking, unset cloexec) but musn't call
to methods that involve the evloop and would mess with the parent
evloop.