Fix: remember maximum index in arena to avoid OOM
We iterate the arena to re-register the fds after fork, but we iterate
up to the total capacity, which is the hard limit of fds the process
open, which happens to be unlimited on darwin (and thus capped to
Int32::MAX) which means we iterate a billion entries *and* the kernel
eventually tries to allocate 96GB of RAM and we each an OOM error.

This didn't reproduce on linux or freebsd that have a reasonable hard
limit (around 1 million) by default and only allocated 40MB. They still
wasted some memory.