Avoid unwinding the stack on hot path in method call lookups (#15002)
The method lookup code uses exceptions to retry lookups using
auto-casting. This is effectively using an exception for execution
control, which is not what they are intended for. On `raise`, Crystal
will try to unwind the call stack and save it to be able to report the
original place where the exception was thrown, and this is a very
expensive operation. To avoid that, we initialize the callstack of the
special `RetryLookupWithLiterals` exception class always with the same
fixed value.