Add Arena::Index to abstract the generation index
Arena::Index now abstracts the underlying i64 representation of the
generation index and allows to access the actual index and generation
number independently.

Arena::Index is now the sole acceptable index after the object is
allocated (be it get or free).

Also fixes an issue in kqueue/evloop for 32-bit systems, where udata can
only be a 32 bit number (oops). Since we already have the fd/index as
the kevent ident, we deconstruct the index to only pass the genration
number, then reconstruct the index from the ident+udata.