Enable runners from `runs-on.com` for Aarch64 CI (#15007)
https://runs-on.com is a service that provisions machines on AWS on demand to run workflow jobs.
It triggers when a job is tagged as `runs-on` and you can configure what kind of machine you like this to run on. The machine specification could likely use some fine tuning (we can use other instance types, and theoretically 8GB should be sufficient). But that'll be a follow-up. For now we know that this works.
We expect this to be more price-efficient setup than renting a fixed server or a CI runner service.

This patch also includes an update to the latest base image. The old arm base images were using Crystal 1.0 and some outdated libraries which caused problems on the new runners (#15005). The build images are based on the docker images from [84codes/crystal](https://hub.docker.com/r/84codes/crystal).