Fix: race condition in Waiters#add vs Waiters#consume_each
One thread may close a file descriptor or run the evloop and receive an
error or hup event, which will wakeup all pending fibers. Further
attempts should fail since the fd is closed, but another thread might
be trying to add a waiter, which may go unnoticed by `#consume_each`
that doesn't manipulate the `@ready` variable.

This patch resolves this by distinguishing between ready (we want to
wakeup one waiter) and closed (the fd will never go to block anymore).

Also simplifies the implementation to rely on the lock, instead of the
lock + atomics. We'll see later if the lock is an issue and how to
optimize in that case.