Fix Deadlock with parallel stop-world/fork calls in MT (#15096)
Trying to stop the world from a thread while threads are forking leads
to a deadlock situation in glibc (at least) because we disable the
reception of *all* signals while we fork.

Since the suspend and resume signals are handled directly and not
processed through the event loop (for obvious reasons) we can safely
keep these signals enabled. Apparently it's even safer.