Add location information to more MacroIf related nodes (#15100)
The most significant addition is when you have a `MacroIf` node like:

```cr
{% if 1 == 2 %}
  {{2 * 2}}
{% end %}
```
The `then` block of it is parsed as an `Expressions` that consists of `[MaccroLiteral, MacroExpression, MacroLiteral]`. I added location information to the nested `MacroExpression` so it knows it's on line 2 instead of line 1, which is what you'd get if you did `node.then.location` since that's the location of the first `MacroLiteral`. This is essentially my workaround for https://github.com/crystal-lang/crystal/issues/14884#issuecomment-2423904262.

This PR also handles `Begin` and `MacroIf` nodes when nested in another node.