Arena: allocate in blocks instead of preallocating a single mmap
The main advantage is the ability to grow on demand and avoid a large
single contiguous mmap of memory that needs to preallocates most of the
memory.

We also don't need to remember the maximum reached index anymore, which
is replaced with a `#each_index` method (only needed to allow `fork`).