Pre-compute `String` size after `#chomp()` if possible (#15153)
If a `String` ends with one of the newlines and already has its number of characters known, the returned `String` must have 1 or 2 fewer characters than `self`. This PR avoids the need to re-compute `@length` in the new string again if possible.

This technique is also applicable to many more substring extraction methods in `String`.