Drop Crystal::FiberChannel
Its main purpose was to keep the eventloop from exiting when empty,
which isn't needed anymore (evloop#run can now wait forever).

Since we need to grab the lock to send the fiber, we can just push it to
the runnables deque instead of passing it through a pipe; which didn't
work properly on Windows anyway.

The Windows IOCP eventloop will need a refactor to check on completions
even where there are no queued (timers).