Add ExecutionContext::SingleThreaded scheduler
Introduces the first EC scheduler that runs in a single thread. Uses the
same queues (Runnables, GlobalQueue) as the multi-threaded scheduler
that will come next. The Runnables local queue could be simplified (no
parallel accesses, hence no need for atomics) at the expense of
duplicating the implementation.

The scheduler doesn't need to actively park the thread, since the event
loops always block (when told to), even when they are no events, which
acts as parking the thread.