Use empty prelude for compiler tools specs (#15272)
Apply some tiny refactors to specs that rely on features from the stdlib prelude. These seem to be not particularly relevant for the test scenario because we're testing compiler features, not stdlib implementations.

These changes allow using the compiler with the empty prelude which results in significantly less compilation effort.

The spec runtime for these tests reduces from 3:06 minutes to 13.81 seconds (over 90% reduction! 🎉🚀)