Fix `tool implementations` to handle gracefully a def with missing location (#15273)
Without this change, `crystal tool implementations` would crash when a target def has no location.

The example uses an enum type's `.new` method which is defined by the compiler without a location.

This patch fixes the nil error. The result may still not be terribly useful, but it's better than crashing the program.