Use Crystal::PointerLinkedList instead of Deque in Mutex
The flat array also doesn't have much impact since it's never traversed:
we only reach the head or the tail once (to dequeue or dequeue one).

This spares a number of GC allocations since the Deque needs to be
allocated and also a buffer that will have to be reallocated sometimes
(and will only ever grow).