Fix: fix bindings + utimes fallback
The `utimensat` and `futimens` syscalls only appeared in macOS 10.14. To
keep some backward compatibility, the x86_64 bindings are fixed to not
declare them. We can however assume the syscalls to be present when
compiling for aarch64.

`Crystal::System::File.utime` now falls back to `utimes` when `futimens`
isn't defined.

Adds the missing `AT_FDCWD` constant to aarch64-android bindings. Fixes
the DragonflyBSD binding.