Add nanosecond precision to `File.utime` on Unix (#15335)
`Crystal::System::File.utime` uses `LibC.utimensat` when available, otherwise falls back to `utimes`.

The `utimensat` syscall is specified in POSIX.1-2008 and was introduced in the following OS releases:

- Linux 2.6.22, glibc 2.6
- macOS 10.14
- FreeBSD 10.3
- OpenBSD 5.0

To keep some backward compatibility, the x86_64-darwin bindings do not
declare it. We can however assume the syscalls to be present when
compiling for aarch64-darwin.

Co-authored-by: Kubo Takehiro <kubo@jiubao.org>