Reference#exec_recursive[_clone] methods aren't fiber aware
Recursion of the `#inspect` and `#pretty_print` methods is handled by a
global hash, which is a per-thread global, but if any of these methods
yield —which will happen because they're writing to an IO— another
fiber running on the same thread will falsely detect a recursion if it
inspects the same object.

This patch moves `Reference#exec_recursive` methods as instances of
`Fiber`, using a per-fiber hash instead of per-thread/process hash.

References #15088