Corrects lib lookup in case-sensitive OSes
Downcase name of Windows import libraries to avoid duplicates references
(Eg. kernel32 and Kernel32) but to correctly find the libraries on
case-sensitive filesystems, like when cross-compiling from Linux.

While on Windows `Kernel32` and `kernel32` result in the correct library
found for linking, it does not work the same way when cross-compiling.

This also applies to Winsock2 (ws2_32) usage.

The change normalizes on the usage of downcased names for these
libraries (which will have no impact when compiling natively on
Windows).