Fix `Reference#exec_recursive`, `#exec_recursive_clone` to be fiber aware (#15361)
Recursion of the `#inspect`, `#pretty_print` and `#clone` methods is handled
by a per-thread global hash. But if any of these methods
yields —which will happen when writing to an IO— another
fiber running on the same thread will falsely detect a recursion if it
inspects the same object.

This patch moves the backing hashes to instance variables of
`Fiber`, using a per-fiber hash instead of per-thread/process hash.

References #15088