Refactor: add fiber safety to crystal/once
Removes the global mutex that prevents concurrent initializers to run,
while still being able to detect reentrancy, by keeping a list of
pending operations, protected by a global lock that doesn't block other
initializers from running.

Using a linked-list of stack allocated operations may sound inefficient
compared to a `Hash` but there should usually not be many concurrent or
parallel lazy initialization that following a few pointers will matter
much in practice.

This isn't very important for the eager initialization of class vars and
constants that happen sequentially at the start of the program and
before we even start threads, but we plan to reuse the feature to
protect lazy initialization of class variables (i.e. the `class_getter`
and `class_property` macros) when it would matter more.

Introduces the `Crystal::Once` namespace. The initialization entry is
now `Crystal::Once.init` instead of `Crystal.once_init`.

Co-authored-by: David Keller <davidkeller@tuta.io>