Disable directory path redirect when `directory_listing=false` (#15393)
With `directory_listing: false` the handler returns 404 Not Found when requesting a directory with the trailing slash (`/foo/`). But, if the trailing slash is missing (`/foo`), it redirects to `/foo/` with a 302 Found response. This leaks information about the existence of directories and can lead to redirect loops if a reverse proxy tries to remove trailing slashes.

Co-authored-by: Julien Portalier <julien@portalier.com>