RFC 2: initialize `Fiber` with an explicit stack
`Fiber::ExecutionContext#spawn` takes advantage of this to take a stack
from the execution context's stack pool, instead of the current
context's stack pool. Cross context spawns thus won't take a stack from
context A but release the stack in context B (where it actually ran)
which would prevent stack recycling.

This also permits to create lots of Fiber instances in specs using a
minimal fake stack, instead of requesting 8MB of virtual memory for each
fiber, despite the fibers never running... which is leaking memory
because we only release the stacks when the fiber terminates.