Add @[Linkage("value")] to fun
The annotation can be specified on fun definitions (with a body).
Incidently it can also be specified to extern fun in lib declarations
(without a body) in which case LLVM may only accept a few values
(external, weak).

The linkage is only set during codegen when `--single-module` is set,
since otherwise it would prevent any module from calling funs that are
always declared in the `main` module.