Move shadow space reservation to x86_64 makecontext
The shadow space (or home space) is a requirement of the x64 call
convention: we must reserve 32 bytes before the return address on the
stack. It only applies to x64, not to arm64 for example.

We don't have to deal with this when creating the stack. This is a
consideration for each individual `makecontext` to handle.