Fix `Tuple#to_a(&)` for arbitrary block output type (#15431)
`Tuple#to_a(&)` is an optimized implementation of `Enumerable#to_a(&)` (overriding `Indexable#to_a(&)`).

But the array is wrongly typed. It is impossible to map to a type that's not in the tuple:

```cr
{1}.to_a(&.to_s) # Error: expected argument #2 to 'Pointer(Int32)#[]=' to be Int32, not String
```

Like the original implementations, `Tuple#to_a` must create an array that can hold the output type of the block because that's what it collects. It has nothing to do with the type of the tuple.