Set linkage of `__crystal_*` funs to internal when compiling to single module
Unlike the usual fun that create a non mangled symbols, and is usually
expected to have external linkage, the `__crystal_*` funs are an
internal implementation detail to abstract calls to stdlib from the
codegen pass (e.g. `__crystal_once`, `__crystal_raise`, ...).

We should compile them to be internal when compiling to a single module.