Simplify `Enumerable#to_a` (#15432)
The implementation of `#to_a` is identical to the yielding variant `#to_a(&)`. We can simply delegate to `#to_a(&)`.
As a bonus, inheriting types which override `#to_a(&)` with an optimized implementation will implicitly use the same optimization for `#to_a` as well. This means we can drop `Indexable#to_a`.

Inheriting types `Tuple` and `Hash` already explicitly delegate `#to_a` to their override implementations of `#to_a(&)`.
We keep these overrides of `#to_a` because they augment the documentation.
But we replace the method bodies with `super` to make it clear that the behaviour is inherited and the `def` only provides documentation.