Move shadow space reservation to x86_64 `makecontext` (#15434)
The shadow space (or home space) is a requirement of the x64 call
convention: we must reserve 32 bytes before the return address on the
stack. It only applies to x64, not to arm64 for example.

We don't have to deal with this when creating the stack. This is a
consideration for each individual `makecontext` to handle.

### Explanation

The old version for win32:

1. aligned the ptr to 16 bytes;
2. decremented the ptr by 6 pointer-size (8 bytes): 1 to move it into the
    stack area (8 bytes), 4 for the shadow space (32 bytes), and 1 last to 
    keep the 16-bytes alignment (another 8 bytes);
3. aligned the ptr to 16 bytes (not needed).

The new version (actually the old non-win32 version):

1. decrements the ptr by 1 pointer-size to move it back _into_ the stack area;
2. aligns the ptr to 16 bytes.

In both versions `stack_ptr` point to the ~~last~~ first (the stack grows 
down) addressable `Void*` in the stack and is aligned to 16 bytes.

Then the new x86_64-microsoft variant of `makecontext` reserves 32 bytes
of shadow space (still aligned to 16 bytes) and the rest of `makecontext` is
still aligned to 16 bytes.

### References

- [Microsoft: x64 call convention](https://learn.microsoft.com/en-us/cpp/build/x64-calling-convention?view=msvc-170)
- [Shadow Space and Stack Alignment (x64)](https://open-advanced-windows-exploitati.gitbook.io/open-advanced-windows-exploitation/custom-shellcode/64-bit-architecture/calling-conventions#shadow-space-and-stack-alignment)
- [Microsoft: arm64 call convention](https://learn.microsoft.com/en-us/cpp/build/arm64-windows-abi-conventions?view=msvc-170)