Set linkage of `__crystal_*` funs to internal (#15439)
Unlike the usual `fun` for which we might expect to have external linkage,
the `__crystal_*` funs are an internal implementation detail to abstract calls
to stdlib from the codegen pass (e.g. `__crystal_once`, `__crystal_raise`, ...).

When compiling to a single module, we can mark them to have internal
linkage, so they're only considered inside the object file. We already do
that for all globals and `def`.