Extract forward compatiblity checks and run on nightly schedule (#15437)
Running compatibility checks with every Crystal release between the oldest forward-compatible (1.0.0) and latest is quite expensive. It currently takes 9:20 hours (14 * 0:35).
And they're not particularly useful. Checks for each individual release would just help discover an odd compatibility issue that only exists on a specific Crystal version. That's is not very common. In case there are any compatibility issues, they'll most likely affect the oldest release (1.0.0) as well. Testing for that should suffice to identify the vast majority of issues.
So we're pulling them out of the pre-merge checks and only run them on a schedule on master (or on demand).
This helps reduce unnecessary CI usage.