Add `Call.new` constructor overload without `obj` parameter (#15441)
Many calls don't have a receiver and so there are a lot of `Call.new` instantiations without a receiver object, which means the first argument `obj` is `nil`.

This patch introduces two new constructor overloads which lack the `obj` parameter. This removes a lot of noise at hundreds of call sites (especially in `parser_spec.cr`).